--- cmd/uplink/cmd/transfer.go
@@ -1,182 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package cmd
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"os"
-	"strings"
-	"time"
-
-	progressbar "github.com/cheggaaa/pb"
-	"github.com/zeebo/errs"
-
-	"storj.io/storj/internal/fpath"
-	"storj.io/storj/pkg/storage/streams"
-	"storj.io/storj/pkg/storj"
-	"storj.io/storj/pkg/stream"
-	"storj.io/storj/pkg/utils"
-)
-
-// upload transfers src from local machine to s3 compatible object dst
-func upload(ctx context.Context, src fpath.FPath, dst fpath.FPath, showProgress bool) (err error) {
-	if !src.IsLocal() {
-		return fmt.Errorf("source must be local path: %s", src)
-	}
-
-	if dst.IsLocal() {
-		return fmt.Errorf("destination must be Storj URL: %s", dst)
-	}
-
-	var expiration time.Time
-	if *expires != "" {
-		expiration, err = time.Parse(time.RFC3339, *expires)
-		if err != nil {
-			return err
-		}
-		if expiration.Before(time.Now()) {
-			return fmt.Errorf("Invalid expiration date: (%s) has already passed", *expires)
-		}
-	}
-
-	// if object name not specified, default to filename
-	if strings.HasSuffix(dst.String(), "/") || dst.Path() == "" {
-		dst = dst.Join(src.Base())
-	}
-
-	var file *os.File
-	if src.Base() == "-" {
-		file = os.Stdin
-	} else {
-		file, err = os.Open(src.Path())
-		if err != nil {
-			return err
-		}
-		defer func() { err = errs.Combine(err, file.Close()) }()
-	}
-
-	fileInfo, err := file.Stat()
-	if err != nil {
-		return err
-	}
-
-	if fileInfo.IsDir() {
-		return fmt.Errorf("source cannot be a directory: %s", src)
-	}
-
-	metainfo, streams, err := cfg.Metainfo(ctx)
-	if err != nil {
-		return err
-	}
-
-	createInfo := storj.CreateObject{
-		RedundancyScheme: cfg.GetRedundancyScheme(),
-		EncryptionScheme: cfg.GetEncryptionScheme(),
-		Expires:          expiration.UTC(),
-	}
-	obj, err := metainfo.CreateObject(ctx, dst.Bucket(), dst.Path(), &createInfo)
-	if err != nil {
-		return convertError(err, dst)
-	}
-
-	reader := io.Reader(file)
-	var bar *progressbar.ProgressBar
-	if showProgress {
-		bar = progressbar.New(int(fileInfo.Size())).SetUnits(progressbar.U_BYTES)
-		bar.Start()
-		reader = bar.NewProxyReader(reader)
-	}
-
-	err = uploadStream(ctx, streams, obj, reader)
-	if err != nil {
-		return err
-	}
-
-	if bar != nil {
-		bar.Finish()
-	}
-
-	fmt.Printf("Created %s\n", dst.String())
-
-	return nil
-}
-
-func uploadStream(ctx context.Context, streams streams.Store, mutableObject storj.MutableObject, reader io.Reader) error {
-	mutableStream, err := mutableObject.CreateStream(ctx)
-	if err != nil {
-		return err
-	}
-
-	upload := stream.NewUpload(ctx, mutableStream, streams)
-
-	_, err = io.Copy(upload, reader)
-
-	return utils.CombineErrors(err, upload.Close())
-}
-
-// download transfers s3 compatible object src to dst on local machine
-func download(ctx context.Context, src fpath.FPath, dst fpath.FPath, showProgress bool) (err error) {
-	if src.IsLocal() {
-		return fmt.Errorf("source must be Storj URL: %s", src)
-	}
-
-	if !dst.IsLocal() {
-		return fmt.Errorf("destination must be local path: %s", dst)
-	}
-
-	metainfo, streams, err := cfg.Metainfo(ctx)
-	if err != nil {
-		return err
-	}
-
-	readOnlyStream, err := metainfo.GetObjectStream(ctx, src.Bucket(), src.Path())
-	if err != nil {
-		return convertError(err, src)
-	}
-
-	download := stream.NewDownload(ctx, readOnlyStream, streams)
-	defer func() { err = errs.Combine(err, download.Close()) }()
-
-	var bar *progressbar.ProgressBar
-	var reader io.Reader
-	if showProgress {
-		bar = progressbar.New(int(readOnlyStream.Info().Size)).SetUnits(progressbar.U_BYTES)
-		bar.Start()
-		reader = bar.NewProxyReader(download)
-	} else {
-		reader = download
-	}
-
-	if fileInfo, err := os.Stat(dst.Path()); err == nil && fileInfo.IsDir() {
-		dst = dst.Join((src.Base()))
-	}
-
-	var file *os.File
-	if dst.Base() == "-" {
-		file = os.Stdout
-	} else {
-		file, err = os.Create(dst.Path())
-		if err != nil {
-			return err
-		}
-		defer func() { err = errs.Combine(err, file.Close()) }()
-	}
-
-	_, err = io.Copy(file, reader)
-	if err != nil {
-		return err
-	}
-
-	if bar != nil {
-		bar.Finish()
-	}
-
-	if dst.Base() != "-" {
-		fmt.Printf("Downloaded %s to %s\n", src.String(), dst.String())
-	}
-
-	return nil
-}
